# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/subscriptions": {
            "get": {
                "description": "Если передать user_id — вернёт только подписки этого пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Список всех подписок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID пользователя для фильтрации",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_maaw77_effm_internal_dto.SubscriptionResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Одна запись = оплата за один месяц. Дубли по пользователю + сервису + месяцу запрещены",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Создать запись о подписке за конкретный месяц",
                "parameters": [
                    {
                        "description": "Данные подписки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_maaw77_effm_internal_dto.CreateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "UUID созданной записи",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "allOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "подписка уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/subscriptions/total": {
            "get": {
                "description": "Период задаётся включительно. Поддерживает фильтры по пользователю и сервису",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Подсчёт общей стоимости подписок за период",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2025-01",
                        "description": "Начало периода (YYYY-MM)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2025-12",
                        "description": "Конец периода (YYYY-MM)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
                        "description": "UUID пользователя",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название сервиса",
                        "name": "service",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_maaw77_effm_internal_dto.TotalCostResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/subscriptions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получить подписку по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_maaw77_effm_internal_dto.SubscriptionResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Можно менять только название сервиса и/или цену. Месяц и год — неизменяемые",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Частично обновить подписку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_maaw77_effm_internal_dto.UpdateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "allOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "status": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "subscriptions"
                ],
                "summary": "Удалить подписку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID подписки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "allOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "status": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_maaw77_effm_internal_dto.CreateSubscriptionRequest": {
            "description": "Данные для создания записи об оплате подписки. Дата в формате YYYY-MM (например, 2025-07)",
            "type": "object",
            "required": [
                "price",
                "service_name",
                "user_id",
                "year_month"
            ],
            "properties": {
                "price": {
                    "description": "Стоимость за месяц в рублях",
                    "type": "integer",
                    "example": 499
                },
                "service_name": {
                    "description": "Название сервиса",
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "user_id": {
                    "description": "UUID пользователя",
                    "type": "string",
                    "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                },
                "year_month": {
                    "description": "Год и месяц в формате YYYY-MM",
                    "type": "string",
                    "example": "2025-07"
                }
            }
        },
        "github_com_maaw77_effm_internal_dto.SubscriptionResponse": {
            "description": "Полная информация о записи подписки с датой создания/обновления",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Время создания",
                    "type": "string",
                    "example": "2025-07-15T10:23:45Z"
                },
                "id": {
                    "description": "UUID записи",
                    "type": "string",
                    "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                },
                "price": {
                    "description": "Стоимость за месяц",
                    "type": "integer",
                    "example": 499
                },
                "service_name": {
                    "description": "Название сервиса",
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "updated_at": {
                    "description": "Время последнего обновления",
                    "type": "string",
                    "example": "2025-07-15T10:23:45Z"
                },
                "user_id": {
                    "description": "UUID пользователя",
                    "type": "string",
                    "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                },
                "year_month": {
                    "description": "Год и месяц в формате YYYY-MM",
                    "type": "string",
                    "example": "2025-07"
                }
            }
        },
        "github_com_maaw77_effm_internal_dto.TotalCostResponse": {
            "description": "Суммарная стоимость всех подходящих подписок за выбранный период",
            "type": "object",
            "properties": {
                "total": {
                    "description": "Итоговая сумма в рублях",
                    "type": "integer",
                    "example": 8599
                }
            }
        },
        "github_com_maaw77_effm_internal_dto.UpdateSubscriptionRequest": {
            "description": "Можно обновить название сервиса и/или цену. Поля опциональные",
            "type": "object",
            "properties": {
                "price": {
                    "description": "Новая цена (если 0 — будет проигнорировано)",
                    "type": "integer",
                    "example": 599
                },
                "service_name": {
                    "description": "Новое название сервиса",
                    "type": "string",
                    "example": "Yandex Plus Premium"
                }
            }
        }
    }
}